+++ app/controllers/paper_tracker_controller.rb
@@ -1,3 +1,4 @@
+# Returns papers being searched for by admins for display in Paper Tracker
 class PaperTrackerController < ApplicationController
   before_action :authenticate_user!
 
@@ -5,11 +6,26 @@ class PaperTrackerController < ApplicationController
 
   def index
     # show all papers that user is connected to across all journals
-    papers = papers_submitted.where(journal_id: journal_ids)
-    respond_with papers, each_serializer: PaperTrackerSerializer, root: 'papers'
+    papers = papers_submitted.where(journal_id: journal_ids).page(page)
+    respond_with papers,
+                 each_serializer: PaperTrackerSerializer,
+                 root: 'papers',
+                 meta: metadata(papers)
   end
 
   private
+
+  def page
+    params[:page].present? ? params[:page] : 1
+  end
+
+  def metadata(papers)
+    {
+      totalCount: papers.total_count, # needs to be called on relation.page
+      perPage: Kaminari.config.default_per_page,
+      page: page
+    }
+  end
 
   def journal_ids
     current_user.old_roles.pluck(:journal_id).uniq